refactor: dto 사용하지 않는 클래스 제거 및 패키지 정리
