+++ back-end/src/main/java/admin/gpuserver/application/GpuServerService.java
@@ -8,10 +8,10 @@ import admin.gpuserver.domain.repository.DeleteHistoryRepository;
 import admin.gpuserver.domain.repository.GpuBoardRepository;
 import admin.gpuserver.domain.repository.GpuServerRepository;
 import admin.gpuserver.domain.repository.LabRepository;
-import admin.gpuserver.dto.GpuServerNameUpdateRequest;
-import admin.gpuserver.dto.GpuServerRequest;
-import admin.gpuserver.dto.GpuServerResponse;
-import admin.gpuserver.dto.GpuServerResponses;
+import admin.gpuserver.dto.request.GpuServerNameUpdateRequest;
+import admin.gpuserver.dto.request.GpuServerRequest;
+import admin.gpuserver.dto.response.GpuServerResponse;
+import admin.gpuserver.dto.response.GpuServerResponses;
 import admin.gpuserver.exception.GpuServerServiceException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
