--- back-end/src/main/java/admin/gpuserver/dto/JobDtos.java
@@ -1,19 +1,0 @@
-package admin.gpuserver.dto;
-
-public class JobDtos {
-    private final RunningJobDto runningJobDto;
-    private final WaitingJobDtos waitingJobDtos;
-
-    public JobDtos(RunningJobDto runningJobDto, WaitingJobDtos waitingJobDtos) {
-        this.runningJobDto = runningJobDto;
-        this.waitingJobDtos = waitingJobDtos;
-    }
-
-    public RunningJobDto getRunningJobDto() {
-        return runningJobDto;
-    }
-
-    public WaitingJobDtos getWaitingJobDtos() {
-        return waitingJobDtos;
-    }
-}
