--- back-end/src/main/java/admin/gpuserver/dto/WaitingJobDto.java
@@ -1,19 +1,0 @@
-package admin.gpuserver.dto;
-
-public class WaitingJobDto {
-    private final String name;
-    private final String expectedRunningTime;
-
-    public WaitingJobDto(String name, String expectedRunningTime) {
-        this.name = name;
-        this.expectedRunningTime = expectedRunningTime;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getExpectedRunningTime() {
-        return expectedRunningTime;
-    }
-}
