+++ back-end/src/main/java/admin/gpuserver/dto/request/GpuBoardRequest.java
@@ -1,4 +1,4 @@
-package admin.gpuserver.dto;
+package admin.gpuserver.dto.request;
 
 public class GpuBoardRequest {
     private String modelName;
