+++ back-end/src/main/java/admin/gpuserver/dto/response/GpuResponses.java
@@ -1,4 +1,4 @@
-package admin.gpuserver.dto;
+package admin.gpuserver.dto.response;
 
 import java.util.List;
