+++ back-end/src/test/java/admin/AcceptanceTest.java
@@ -1,0 +1,4 @@
+package admin;
+
+public class AcceptanceTest {
+}
